--- a/Gemfile
+++ b/Gemfile
@@ -1,6 +1,7 @@
-source 'https://rubygems.org'
-gem 'jekyll', '4.2.0'
-gem 'kramdown-parser-gfm'
-gem 'jekyll-redirect-from'
-gem 'rouge', '3.26.0'
-gem 'liquid-c'
+source "https://rubygems.org"
+gem "jekyll", "4.2.0"
+gem "kramdown-parser-gfm"
+gem "jekyll-redirect-from"
+gem "rouge", "3.26.0"
+gem "liquid-c"
+gem "jekyll-include-cache"