--- a/app/_includes/head.html
+++ b/app/_includes/head.html
@@ -98,7 +98,7 @@ site.title }}{% endcapture %} {% else %} {% capture page_title %}{{ page.name }}
   />
 
 
-  {% if page.path == "_hub/index.html" %}
+  {% if include.is_plugin_hub %}
   <link
     rel="stylesheet"
     href="/assets/hub.css?v={{ site.time | date: '%s' }}"