--- a/app/_layouts/docs-v2.html
+++ b/app/_layouts/docs-v2.html
@@ -4,38 +4,42 @@ sitemap: true
 id: documentation
 ---
 
-{% assign latest_page_url = page.url | replace: page.kong_version, page.kong_latest.release %}
-{% capture latest_page_exists %}{% page_exists {{latest_page_url}} %}{% endcapture %}
+{% assign latest_page_url = page.url | replace: page.kong_version,
+page.kong_latest.release %}
+{% capture latest_page_exists %}{% page_exists {{latest_page_url}} %}
+{% endcapture %}
 
-<div class="page v2 {{ page.class }}">
+<div class="page v2 {{ page.class }}" data-url={{ page.url }}>
 
   <div class="container">
-    {% include docs-sidebar.html %}
+    {% include_cached docs-sidebar.html kong_version=page.kong_version edition=page.edition %}
 
     {% if page.toc == true or page.toc != false and filename != "index.md" and page.is_homepage != true %}
-    <aside class="docs-toc">
-      <i class="fa fa-times close-sidebar"></i>
-      <i class="fa fa-chevron-right collapse-toc"></i>
-      <i class="far fa-list-alt expand-toc"></i>
-      <div class="github-links">
-        <p>
+      <aside class="docs-toc">
+        <i class="fa fa-times close-sidebar"></i>
+        <i class="fa fa-chevron-right collapse-toc"></i>
+        <i class="far fa-list-alt expand-toc"></i>
+        <div class="github-links">
+          <p>
             <a href="https://github.com/Kong/docs.konghq.com/edit/main/app/{{page.path}}" target="_blank">
-            <img src="/assets/images/logos/logo-github.svg" alt="github-edit-page"/>Edit this page</a>
+              <img src="/assets/images/logos/logo-github.svg" alt="github-edit-page"/>Edit this page</a>
           </p>
           <p>
             <a href="https://github.com/Kong/docs.konghq.com/issues" target="_blank">
-            <img src="/assets/images/icons/documentation/icn-monitoring-black.svg" alt="report-issue"/>Report an issue</a>
+              <img src="/assets/images/icons/documentation/icn-monitoring-black.svg" alt="report-issue"/>Report an issue</a>
           </p>
-      </div>
-      <div class="docs-toc-title"><a href="#">On this page</a></div>
-      {% include toc.html html=content anchor_class="scroll-to" h_min=2 h_max=3 %}
-    </aside>
+        </div>
+        <div class="docs-toc-title">
+          <a href="#">On this page</a>
+        </div>
+        {% include_cached toc.html html=content anchor_class="scroll-to" h_min=2 h_max=3 %}
+      </aside>
     {% endif %}
 
     <div class="page-content-container v2 {% if page.no_version == true and page.has_version == false and page.nav_items == nil %}no-sidebar{% endif %}" id="documentation">
       <div class="toggles {% if page.no_version == true and page.has_version == false and page.nav_items == nil %}no-sidebar{% endif %}">
         {% if page.toc == true or page.toc != false and filename != "index.md" and page.is_homepage != true and page.is_homepage != true %}
-        <i class="far fa-list-alt toc-sidebar-toggle"></i>
+          <i class="far fa-list-alt toc-sidebar-toggle"></i>
         {% endif %}
       </div>
 
@@ -44,91 +48,92 @@ id: documentation
           <div class="content-header">
             <div class="content-edition">
               {% if page.edition == 'enterprise' %}{{site.ee_product_name}}
-              {% elsif page.edition == 'konnect' %}{{site.konnect_saas}}
-              {% elsif page.edition == 'mesh' %}{{site.mesh_product_name}}
-              {% elsif page.edition == 'deck' %}decK
+                {% elsif page.edition == 'konnect' %}{{site.konnect_saas}}
+                {% elsif page.edition == 'mesh' %}{{site.mesh_product_name}}
+                {% elsif page.edition == 'deck' %}decK
               {% elsif page.edition == 'kubernetes-ingress-controller' %}{{site.kic_product_name}}
-              {% elsif page.edition == 'gateway-oss' %}{{site.ce_product_name}}
-              {% elsif page.edition == 'getting-started-guide' %}Getting Started with Kong Gateway
+                {% elsif page.edition == 'gateway-oss' %}{{site.ce_product_name}}
+                {% elsif page.edition == 'getting-started-guide' %}Getting Started with Kong Gateway
               {% elsif page.edition == 'contributing' %}Contribution guidelines
               {% endif %}
             </div>
             {% unless page.no_version == true %}
-            <div class="content-version">
-              {{page.kong_version}}
-              {% if page.kong_version == page.kong_latest.release %}
-                <span>(latest)</span>
-              {% endif %}
-            </div>
+              <div class="content-version">
+                {{page.kong_version}}
+                {% if page.kong_version == page.kong_latest.release %}
+                  <span>(latest)</span>
+                {% endif %}
+              </div>
             {% endunless %}
             {% unless page.no_breadcrumbs %}
-            {% include breadcrumbs.html %}
+              {% include_cached breadcrumbs.html edition=page.edition dir=page.dir%}
             {% endunless %}
           </div>
           {% if page.edition == 'gateway-oss' %}
-          <div class="alert alert-ee">
-            Maybe you were looking for the <a href="/enterprise"><strong>{{site.ee_product_name}} documentation</strong></a>
+            <div class="alert alert-ee">
+            Maybe you were looking for the <a href="/enterprise">
+                <strong>{{site.ee_product_name}} documentation</strong>
+              </a>
             instead?
           </div>
           {% endif %}
           {% if page.kong_version != page.kong_latest.release and page.no_version != true %}
-          <div class="alert alert-warning">
+            <div class="alert alert-warning">
             You are browsing documentation for an outdated version. See the
             <a href="{% if latest_page_exists == 'true' %}{{latest_page_url}}{% elsif page.edition == 'gateway-oss'%}/gateway-oss/{% elsif page.edition == 'enterprise'%}/enterprise/{% elsif page.edition == 'mesh'%}/mesh/{% elsif page.edition == 'kubernetes-ingress-controller/'%}/kubernetes-ingress-controller/{% else %}/{% endif %}">
               latest documentation here</a>.
           </div>
-          {% elsif page.beta == true %}
-          <div class="alert alert-ee condensed margin-bigger">
-            <div class="alert-body">
-              <div class="tag">BETA</div>
-              <p>Warning: This feature is released as <a href="/enterprise/{{page.kong_version}}/introduction/key-concepts/#beta">BETA</a> and should not be deployed in a production environment.
+            {% elsif page.beta == true %}
+            <div class="alert alert-ee condensed margin-bigger">
+              <div class="alert-body">
+                <div class="tag">BETA</div>
+                <p>Warning: This feature is released as <a href="/enterprise/{{page.kong_version}}/introduction/key-concepts/#beta">BETA</a> and should not be deployed in a production environment.
             </div>
-          </div>
-          {% endif %}
+              </div>
+            {% endif %}
 
-          <h1 class="page-content-title{%
+            <h1 class="page-content-title{%
             if page.badge %} badge{%
             if page.badge == 'plus' %} plus{% endif %}{%
             if page.badge == 'enterprise' %} enterprise{% endif %}
             {% endif %}"
             >{{page.title | flatify }}</h1>
-          {% if page.enterprise %}
-          {% endif %}
-          {% if page.subtitle %}
-          <h2 class="page-content-subtitle">{{page.subtitle | flatify }}</h2>
-          {% endif %}
+            {% if page.enterprise %}{% endif %}
+            {% if page.subtitle %}
+              <h2 class="page-content-subtitle">{{page.subtitle | flatify }}</h2>
+            {% endif %}
 
-          {{ content }}
+            {{ content }}
 
+          </div>
         </div>
       </div>
     </div>
-  </div>
-  {% if page.is_homepage != true %}
-  <div id="scroll-to-top-button">
-    <i class="fas fa-chevron-up"></i>
-  </div>
-    {% if page.feedback != false %}
-    {% include feedback-widget.html %}
+    {% if page.is_homepage != true %}
+      <div id="scroll-to-top-button">
+        <i class="fas fa-chevron-up"></i>
+      </div>
+      {% if page.feedback != false %}
+        {% include_cached feedback-widget.html %}
+      {% endif %}
     {% endif %}
-  {% endif %}
-</div>
+  </div>
 
-{% include image-modal.html %}
+  {% include_cached image-modal.html %}
 
-<script type="text/javascript">
-  function toggleContentHeader () {
-    var w = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
-    if (w <= 800 && $(window).scrollTop() < 8) {
-      $('.content').toggleClass('scrolled', false);
-      $('.content-header').toggleClass('scrolled', false);
-      $('.content-version').show(250);
-    } else {
-      $('.content').toggleClass('scrolled', true);
-      $('.content-header').toggleClass('scrolled', true);
-      $('.content-version').hide(250);
+  <script type="text/javascript">
+    function toggleContentHeader() {
+      var w = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
+      if (w <= 800 && $(window).scrollTop() < 8) {
+        $('.content').toggleClass('scrolled', false);
+        $('.content-header').toggleClass('scrolled', false);
+        $('.content-version').show(250);
+      } else {
+        $('.content').toggleClass('scrolled', true);
+        $('.content-header').toggleClass('scrolled', true);
+        $('.content-version').hide(250);
+      }
     }
-  }
 
-  window.addEventListener('scroll', toggleContentHeader);
-</script>
+    window.addEventListener('scroll', toggleContentHeader);
+  </script>