--- a/app/_layouts/extension.html
+++ b/app/_layouts/extension.html
@@ -228,7 +228,7 @@ breadcrumbs:
         {% endif %}
 
         {% if page.params.examples != false %}
-          {% include hub-examples.html %}
+          {% include_cached hub-examples.html %}
         {% endif %}
 
         <h3 id="parameters">Parameters</h3>
@@ -273,7 +273,7 @@ breadcrumbs:
       </div>
       {% unless page.toc == false %}
         <nav>
-          {% include toc.html html=full_content h_min=2 h_max=3 sanitize=true %}
+          {% include_cached toc.html html=full_content h_min=2 h_max=3 sanitize=true %}
         </nav>
       {% endunless %}