--- a/app/_layouts/install.html
+++ b/app/_layouts/install.html
@@ -59,7 +59,7 @@ layout: default
           {% if page.toc == true or
             page.toc != false and filename != "index.md" %}
             <h2 id="table-of-contents">Table of Contents</h2>
-            {% include toc.html html=content class="inline-toc" anchor_class="scroll-to" h_max=2 %}
+            {% include_cached toc.html html=content class="inline-toc" anchor_class="scroll-to" h_max=2 %}
             {% endif %}
         <div class="content">
           {{ content }}