### Review

@lena-larionova 

### Summary
Add a liquid parser written in C to reduce the build time. This is a quick win to speed up the build.

It also adds the `jekyll-include-cache` gem which speeds up `include` usage. We use `include` heavily when generating navigation.

Anecdotal evidence shows a ~70% improvement on my local machine and a ~60% increase on netlify.

## Locally

Without Cache, without liquid-c

* 103.905 seconds
* 72.69 seconds
* 70.078 seconds
* 70.817 seconds

Without Cache, with liquid-c

* 89.737 seconds
* 73.613 seconds
* 69.113 seconds
* 69.413 seconds

With Cache, with liquid-c

* 54.5 seconds
* 26.62 seconds
* 27.44 seconds
* 27.026 seconds

## On Netlify

Before change:

* 142.009 seconds
* 153.586 seconds
* 138.788 seconds

After change (liquid-c):

* 115.321 seconds

After change (liquid-c and include_cached)

* 59.0772

### Reason
Less time waiting 🥳

### Testing
Spot check various pages. Please pay attention to things that change per page such as navigation, breadcrumbs, table of contents etc

<!-- (Optional) Include link to topic in Netlify preview after it's generated 
(around 10mins after PR is created) -->