@lena-larionova 2021-07-08
Just found two more places where caching is still not working quite right:

* Active nav state in versioned guides. The active state is saved for the first topic you click on and then persists no matter where you switch to after.
* Breadcrumbs will probably also need the version passed to the include, as they're currently also caching incorrectly.