@mheap 2021-07-09
Both of those issues should now be resolved.

* Active navigation is now added using javascript rather than the `active` class in Jekyll. This is a reasonable solution as we already use JS to expand the accordion for active navigation so there are no new dependencies. If this solution is accepted we can simplify the templates to remove the active logic (which I have not done yet)
* Breadcrumbs are now passed `page.edition` and `page.dir` which is enough to make each breadcrumb rendering unique. They still use the `page.*` variables internally and there is an opportunity to convert those to `include.*` in the future should this PR be accepted